+++ src/components/meal-card/MealCard.tsx
@@ -1,16 +1,15 @@
 import _ from "lodash";
-import { Fragment, useContext } from "react";
+import { useContext } from "react";
 import { Card, ListGroup } from "react-bootstrap";
-import { Action, deleteMealAction, enterFoodEditModeAction, enterMealEditModelAction, exitFoodEditModeAction, exitMealEditModeAction } from "../../actions";
+import { Action, deleteMealAction, enterMealEditModelAction, exitMealEditModeAction } from "../../actions";
 import { MealState } from "../../model/AppState";
 import { DeleteButton } from "../DeleteButton";
 import { EditModeButton } from "../EditModeButton";
-import { FoodItem } from "../FoodItem";
 import { MealDispatch } from "../MealDispatch";
 import { MealSummary } from "../MealSummary";
 import { AddButtonGroupItem } from "./AddButtonGroupItem";
 import { AddFoodFormGroupItem } from "./AddFoodFormGroupItem";
-import { UpdateFoodFormGroupItem } from "./UpdateFoodFormGroupItem";
+import { FoodGroupItems } from "./FoodGroupItems";
 
 interface Props {
   state: MealState;
@@ -31,28 +30,6 @@ export const MealCard = (props: Props) => {
       : dispatch(enterMealEditModelAction(mealIndex));
   }
 
-  const foodItems = foods.map((food, index) => {
-    const toggleFoodEditMode = () => {
-      index === foodEditIndex
-        ? dispatch(exitFoodEditModeAction(mealIndex))
-        : dispatch(enterFoodEditModeAction(mealIndex, index));
-    }
-    return (
-      <Fragment>
-        <ListGroup.Item key={index} className="d-flex align-items-center">
-          <FoodItem food={food} />
-          {editState === 'edit'
-            && <EditModeButton editMode={index === foodEditIndex} onClick={toggleFoodEditMode} />}
-        </ListGroup.Item>
-        {editState === 'edit'
-          && index === foodEditIndex
-          &&
-          <UpdateFoodFormGroupItem food={food} mealIndex={mealIndex} foodIndex={index} />
-        }
-      </Fragment>
-    )
-  });
-
   return (
     <Card className="mt-1">
       <Card.Header className="d-flex align-items-center">
@@ -66,7 +43,7 @@ export const MealCard = (props: Props) => {
       </Card.Header>
 
       <ListGroup>
-        {foodItems}
+        <FoodGroupItems foods={foods} mealIndex={mealIndex} foodEditIndex={foodEditIndex} editState={editState} />
         {editState === 'edit' && <AddButtonGroupItem mealIndex={mealIndex} />}
         {editState === 'add' && <AddFoodFormGroupItem mealIndex={mealIndex} />}
       </ListGroup>
